requirements.txt
asgiref==3.8.1
django==5.1
django-taggit==6.0.0
markdown==3.7
psycopg==3.1.18
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.1
typing-extensions==4.12.2
